Pulled out a TraitToEmbedTransformer from TraitsModule
