+++ Orabot.Core/Modules/OpenRaTraitsModule.cs
@@ -1,36 +1,31 @@
 ﻿using System.Threading.Tasks;
-using Discord;
 using Discord.Commands;
 using Microsoft.Extensions.Configuration;
-using RestSharp;
+using Orabot.Core.Transformers.DocumentationToEmbedTransformers;
 
 namespace Orabot.Core.Modules
 {
 	public class OpenRaTraitsModule : ModuleBase<SocketCommandContext>
 	{
-		private readonly IRestClient _restClient;
-		private readonly string _openRaIconUrl;
 		private readonly string _traitsReleasePageUrl;
 		private readonly string _traitsPlaytestPageUrl;
 		private readonly string _traitsDevelopmentPageUrl;
+		private readonly TraitToEmbedTransformer _traitToEmbedTransformer;
 
-		public OpenRaTraitsModule(IRestClient restClient, IConfiguration configuration)
+		public OpenRaTraitsModule(TraitToEmbedTransformer traitToEmbedTransformer, IConfiguration configuration)
 		{
-			_restClient = restClient;
-
-			_openRaIconUrl = configuration["OpenRaFaviconUrl"];
-
 			var traitsPages = configuration.GetRequiredSection("Traits");
 			_traitsReleasePageUrl = traitsPages["ReleasePageUrl"];
 			_traitsPlaytestPageUrl = traitsPages["PlaytestPageUrl"];
 			_traitsDevelopmentPageUrl = traitsPages["DevelopmentPageUrl"];
+			_traitToEmbedTransformer = traitToEmbedTransformer;
 		}
 
 		[Command("traits")]
 		[Summary("Provides a link to the OpenRA Traits documentation page. Can be used with an optional trait name to link directly.")]
 		public async Task Traits(string traitName = null)
 		{
-			var embed = await BuildTraitsPageEmbed(_traitsReleasePageUrl, traitName);
+			var embed = await _traitToEmbedTransformer.CreateEmbed(_traitsReleasePageUrl, traitName);
 			if (embed != null)
 				await ReplyAsync("", false, embed);
 		}
@@ -39,7 +34,7 @@ namespace Orabot.Core.Modules
 		[Summary("Provides a link to the OpenRA playtest Traits documentation page. Can be used with an optional trait name to link directly.")]
 		public async Task TraitsPt(string traitName = null)
 		{
-			var embed = await BuildTraitsPageEmbed(_traitsPlaytestPageUrl, traitName);
+			var embed = await _traitToEmbedTransformer.CreateEmbed(_traitsPlaytestPageUrl, traitName);
 			if (embed != null)
 				await ReplyAsync("", false, embed);
 		}
@@ -48,45 +43,13 @@ namespace Orabot.Core.Modules
 		[Summary("Provides a link to the OpenRA development Traits documentation page. Can be used with an optional trait name to link directly.")]
 		public async Task TraitsDev(string traitName = null)
 		{
-			var embed = await BuildTraitsPageEmbed(_traitsDevelopmentPageUrl, traitName);
+			var embed = await _traitToEmbedTransformer.CreateEmbed(_traitsDevelopmentPageUrl, traitName);
 			if (embed != null)
 				await ReplyAsync("", false, embed);
 		}
 
 		#region Private methods
 
-		private async Task<bool> CheckTraitExists(string pageUrl, string traitName)
-		{
-			var request = new RestRequest(pageUrl);
-			var response = await _restClient.GetAsync(request);
-			return response.Content?.Contains($"<a href=\"#{traitName.ToLower()}\"") ?? false;
-		}
-
-		private async Task<Embed> BuildTraitsPageEmbed(string pageUrl, string traitName)
-		{
-			if (string.IsNullOrWhiteSpace(pageUrl))
-				return null;
-
-			var hasName = !string.IsNullOrWhiteSpace(traitName);
-			if (hasName)
-				hasName = await CheckTraitExists(pageUrl, traitName);
-
-			var targetUrl = pageUrl + (hasName ? $"#{traitName.ToLower()}" : string.Empty);
-			var embedBuilder = new EmbedBuilder
-			{
-				Author = new EmbedAuthorBuilder
-				{
-					Name = "OpenRA Traits page" + (hasName ? $", trait {traitName}" : string.Empty),
-					Url = targetUrl,
-					IconUrl = _openRaIconUrl
-				},
-				Title = targetUrl,
-				Url = targetUrl,
-				Description = "This documentation is aimed at modders. It displays all traits with default values and developer commentary."
-			};
-
-			return embedBuilder.Build();
-		}
 
 		#endregion
 	}
